:recycle: (pc-14782) extract functions for readability
