+++ src/features/culturalSurvey/helpers/utils.ts
@@ -55,3 +55,23 @@ export const createInitialQuestionsList = (
   })
   return initialQuestionsList
 }
+
+export const addSubQuestionToQuestionsToDisplay = (
+  subQuestionId: CulturalSurveyQuestionEnum,
+  questionsToDisplay: CulturalSurveyQuestionEnum[]
+) => {
+  let updatedQuestionsToDisplay = [...questionsToDisplay]
+  updatedQuestionsToDisplay = updatedQuestionsToDisplay.filter(
+    (questionId) => questionId != subQuestionId
+  )
+  return updatedQuestionsToDisplay
+}
+
+export const removeSubQuestionsToDisplay = (
+  subQuestionId: CulturalSurveyQuestionEnum,
+  questionsToDisplay: CulturalSurveyQuestionEnum[]
+) => {
+  const updatedQuestionsToDisplay = [...questionsToDisplay]
+  updatedQuestionsToDisplay.splice(1, 0, subQuestionId)
+  return updatedQuestionsToDisplay
+}
